Moving expiresOn to AAD TokenResponse child class
